Add a command for starting dev server with automatic cache clearing
This commit adds a new command “fraction server.start” which can be used to start a development server with the given port number. The server is set up to automatically clear the cache when any changes are detected under the “src” folder. This will help ensure that developers are always working with the latest version of the code. Don't use in production.